Refactor `property_tests.rs` into `property_tests` module structure.
